Fix fraxETH Redeem test and move WETH replacement to mocks
